드래곤 스킬 Breath구현,Character 클래스 구조 변경,카메라 회전 구현
드래곤 스킬 Breath구현
- 공격 범위 표시 후 스킬 시전
- 삼각형 내의 플레이어가 있다면 틱데미지를 줌
- 삼각형 판정 함수 구현
- Breath공격범위 삼각형은 GameObject 3개로 범위 변경가능

Character 클래스 구조 변경
- 템플릿 삭제, ENUM을 위한 템플릿이었으나 int로 변경
- Player_Skill에서 Character_Skill로 변경, Character를 상속받은 object는 스킬 사용 가능

카메라 회전 구현
- 플레이어를 중심으로 마우스 드래그시 회전
- 조이스틱을 클릭 시에는 회전에서 제외

+++ Project_C/Assets/Script/Monster/MonsterState/Monster_AttackState.cs
@@ -22,6 +22,6 @@ public class Monster_AttackState : StateBase
     public override void OnStateUpdate()
     {
         if (monster.IsPlayerInAttackRange() == false)
-            monster.ChangeState(MONSTER_STATE.PATROL);
+            monster.ChangeState((int)MONSTER_STATE.PATROL);
     }
 }
