+++ Project_C/Assets/Script/Monster/MonsterState/Monster_PatrolState.cs
@@ -28,7 +28,7 @@ public class Monster_PatrolState : StateBase
         if (monster.IsPlayerInDetectionRange())
         {
             monster.ChageTarget(Shared.GameMgr.PLAYEROBJ);
-            monster.ChangeState(MONSTER_STATE.CHASE);
+            monster.ChangeState((int)MONSTER_STATE.CHASE);
         }
         //Debug.Log("OnIdleUpdate");
     }
